chore(client): 마이페이지 css 수정
